Fix Normalize, Move cloneMesage to runutill
